refactor: merge nav-items + settings-nav-items and deploy-page + deploy-new-page aliases
Consolidate turbopack resolveAlias entries from 21 to 19 by merging
related extension points that share the same domain:

- nav-items + settings-nav-items → nav-config (both pure navigation data)
- deploy-page + deploy-new-page → deploy-pages (same feature, trivial OSS stubs)
