+++ apps/web/next.config.js
@@ -18,18 +18,16 @@ const nextConfig = {
           "@/lib/auth/auth-provider": "@/cloud/auth/cognito-provider",
           "@/lib/auth/auth-server": "@/cloud/auth/cognito-server",
           "@/lib/actions/resolve-user": "@/cloud/auth/resolve-user",
-          "@/lib/nav-items": "@/cloud/nav-items",
+          "@/lib/nav-config": "@/cloud/nav-config",
           "@dashboard/dashboard-sidebar": "@/cloud/dashboard/dashboard-sidebar",
           "@dashboard/dashboard-header": "@/cloud/dashboard/dashboard-header",
           "@/lib/crypto": "@/cloud/kms-crypto",
           "@/lib/gateway-auth": "@/cloud/gateway-auth",
-          "@/lib/settings-nav-items": "@/cloud/settings-nav-items",
+          "@/lib/deploy/deploy-pages": "@/cloud/deploy/deploy-pages",
           "@/lib/auth/login-content": "@/cloud/auth/login-content",
           "@/lib/user-plan": "@/cloud/user-plan",
           "@/lib/auth/session-hooks": "@/cloud/auth/session-hooks",
           "@/lib/deploy/nav-deploy": "@/cloud/deploy/nav-deploy",
-          "@/lib/deploy/deploy-page": "@/cloud/deploy/deploy-page",
-          "@/lib/deploy/deploy-new-page": "@/cloud/deploy/deploy-new-page",
           "@/lib/components/request-app-slot": "@/cloud/apps/request-app-slot",
           "@/lib/actions/agents": "@/cloud/actions/agents",
           "@/lib/actions/rules": "@/cloud/actions/rules",
