--- apps/web/src/lib/deploy/deploy-new-page.tsx
@@ -1,6 +1,0 @@
-import { redirect } from "next/navigation";
-
-// OSS: no deploy wizard — redirect to the main deploy page (cloud CTA)
-export default function DeployNewPage() {
-  redirect("/deploy");
-}
