+++ apps/web/src/lib/nav-config.ts
@@ -1,4 +1,14 @@
-import { User, KeyRound, ShieldCheck } from "lucide-react";
+import {
+  LayoutDashboard,
+  Bot,
+  Shield,
+  Settings,
+  Plug,
+  User,
+  KeyRound,
+  ShieldCheck,
+} from "lucide-react";
+import type { NavItem } from "@/app/(dashboard)/_components/nav-main";
 
 export interface SettingsNavItem {
   title: string;
@@ -10,6 +20,14 @@ export interface SettingsNavSection {
   label: string;
   items: SettingsNavItem[];
 }
+
+export const navItems: NavItem[] = [
+  { title: "Overview", url: "/overview", icon: LayoutDashboard },
+  { title: "Agents", url: "/agents", icon: Bot },
+  { title: "Rules", url: "/rules", icon: Shield },
+  { title: "Connections", url: "/connections", icon: Plug },
+  { title: "Settings", url: "/settings", icon: Settings },
+];
 
 export const settingsSections: SettingsNavSection[] = [
   {
